Some edits
I found some of the introductory pieces a little hard to follow.

The general thrust here is that we want to add "consistency" without
sacrificing "correctness".  But the definitions were a little heavy on
the latter.  I think that we can lean more on the fact that
losing "correctness" is more obviously bad.

I added some paragraph breaks  and fixed a bad reference.