More about verification before use
A lot of the emphasis on the draft is in finding ways to acquire keys
that are consistent.  It might be easier use a simpler protocol for
acquisition, which might make correctness easier to prove, then seek
separate ways of ensuring consistency.

This change does a little restructuring along those lines.  It probably
needs a deeper edit than this, but I thought that I'd try to get the
ideas down.